change current weapon (needs rev, descr)
SDL_CONTROLLER_BUTTON_LEFTSHOULDER works but SDL_CONTROLLER_BUTTON_RIGHTSHOULDER won't work

+++ DoomEngine/DoomEngine/j1Entity.h
@@ -105,8 +105,8 @@ public:
 	POINTING_DIR            pointingDir = POINTING_DIR::LEFT; 
 	POINTING_DIR            defaultPointingDir = POINTING_DIR::LEFT; 
 	fPoint                  lastSpeed = fPoint(0, 0);
-	float					life = (float)LONG_MAX;
-	float					maxLife = (float)LONG_MAX;
+	int				     	life = INT_MAX;
+	int					    maxLife = INT_MAX;
 	float					speed = (float)LONG_MAX;
 	float                   gravityFactor = (float)LONG_MAX; 
 	float                   mass = (float)LONG_MAX;
